Search + Tab working together. TabBarDelegate in MasterView
